requirements.txt
cs50==9.2.2
Flask==2.2.2
Flask-Session==0.4.0
Flask-Mail==0.9.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
psycopg2==2.9.3
python-dateutil==2.8.1
requests==2.26.0
stripe==4.1.0
